% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_variance.R
\name{analyze_variance}
\alias{analyze_variance}
\title{analyze_variance}
\usage{
analyze_variance(linear_model)
}
\arguments{
\item{linear_model}{The linear model to analyze}
}
\value{
The analysis of variance for the linear model
}
\description{
Analyzes the variance of a linear model
}
\examples{
analysis_of_variance <- analyze_variance(lm(iris$Sepal.Length ~ iris$Sepal.Width, data = iris))
}
