% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeVariance.R
\name{analyzeVariance}
\alias{analyzeVariance}
\title{analyzeVariance}
\usage{
analyzeVariance(linear_model)
}
\arguments{
\item{linear_model}{The linear model to analyze}
}
\value{
The analysis of variance for the linear model
}
\description{
Analyzes variance
}
\examples{
the_analysis_of_variance <- analyzeVariance(the_linear_model)
}
