% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_critical_value_Fc.R
\name{calculate_critical_value_Fc}
\alias{calculate_critical_value_Fc}
\title{calculate_critical_value_Fc}
\usage{
calculate_critical_value_Fc(
  significance_level,
  regression_degrees_of_freedom,
  residual_degrees_of_freedom
)
}
\arguments{
\item{The}{residual degrees of freedom}
}
\value{
The critical value F
}
\description{
Calculates critical value Fc of an F distribution with specified significance level, number of regression degrees of freedom, and number of residual degrees of freedom
}
\examples{
critical_value_F <- calculate_critical_value_Fc(0.05, 2, 52)
}
