% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_critical_value_t.R
\name{calculate_critical_value_t}
\alias{calculate_critical_value_t}
\title{calculate_critical_value_t}
\usage{
calculate_critical_value_t(
  significance_level,
  number_of_confidence_intervals,
  residual_degrees_of_freedom
)
}
\arguments{
\item{The}{degrees of freedom of the Student's t distribution}
}
\value{
The critical value t
}
\description{
Calculates critical value t of a Student's t distribution with a provided number of degrees of freedom with provided significance level and number of confidence intervals
}
\examples{
critical_value_t <- calculate_critical_value_t(0.05, 1, 108)
}
