% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_critical_value_zc.R
\name{calculate_critical_value_zc}
\alias{calculate_critical_value_zc}
\title{calculate_critical_value_zc}
\usage{
calculate_critical_value_zc(significance_level, hypothesis_test_is_two_tailed)
}
\arguments{
\item{The}{indicator of whether of not hypothesis test is two tailed}
}
\value{
The critical value zc
}
\description{
Calculates critical value zc of a standard normal distribution with provided significance level and indicator of whether or not hypothesis test is two tailed
}
\examples{
critical_value_zc <- calculate_critical_value_zc(0.05, TRUE)
}
