% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_fraction_of_correct_predictions_for_LR_model.R
\name{calculate_fraction_of_correct_predictions_for_LR_model}
\alias{calculate_fraction_of_correct_predictions_for_LR_model}
\title{calculate_fraction_of_correct_predictions_for_LR_model}
\usage{
calculate_fraction_of_correct_predictions_for_LR_model(LR_model, data_set)
}
\arguments{
\item{LR_model}{The Logistic Regression model for which to calculate the fraction of correct predictions given a data set for which to predict}
}
\value{
fraction_of_correct_predictions
}
\description{
Calculates the fraction of correct predictions for an Logistic Regression model given a data set for which to predict
}
\examples{
fraction_of_correct_predictions <- calculate_fraction_of_correct_predictions_for_LR_model(LR_model = Direction_vs_Lag2, data_set = Weekly_from_2009_to_2010_inclusive)
}
