% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_squared_error.R
\name{calculate_mean_squared_error}
\alias{calculate_mean_squared_error}
\title{calculate_mean_squared_error}
\usage{
calculate_mean_squared_error(
  vector_of_predicted_response_values,
  vector_of_actual_response_values
)
}
\arguments{
\item{model}{The model}
}
\value{
mean_squared_error The mean squared error
}
\description{
Calculates mean squared error
}
\examples{
mean_squared_error <- calculate_mean_squared_error(model)
}
