% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mode.R
\name{calculate_mode}
\alias{calculate_mode}
\title{calculate_mode}
\usage{
calculate_mode(data_for_sample)
}
\arguments{
\item{data_for_sample}{A vector of data for a sample}
}
\value{
The mode
}
\description{
Calculates the mode given data for a sample
}
\examples{
mode <- calculate_mode(iris$Sepal.Length)
}
