% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_p_value_from_t_statistic_and_residual_degrees_of_freedom.R
\name{calculate_p_value_from_t_statistic_and_residual_degrees_of_freedom}
\alias{calculate_p_value_from_t_statistic_and_residual_degrees_of_freedom}
\title{calculate_p_value_from_t_statistic_and_residual_degrees_of_freedom}
\usage{
calculate_p_value_from_t_statistic_and_residual_degrees_of_freedom(
  t_statistic,
  residual_degrees_of_freedom
)
}
\arguments{
\item{The}{residual degrees of freedom}
}
\value{
The p-value corresponding to a provided test statistic t of a Student's t distribution with a provided number of degrees of freedom
}
\description{
Calculates the p-value corresponding to a provided test statistic t of a Student's t distribution with a provided number of degrees of freedom
}
\examples{
p_value <- calculate_p_value_from_t_statistic_and_residual_degrees_of_freedom(1.982, 108)
}
