% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentile.R
\name{calculate_percentile}
\alias{calculate_percentile}
\title{calculate_percentile}
\usage{
calculate_percentile(data_for_sample, value)
}
\arguments{
\item{data_for_sample}{A vector of data for a sample}

\item{value}{A number for which to find the percentile given data for a sample}
}
\value{
A percentile
}
\description{
Calculates the percentile of a value given data for a sample
}
\examples{
percentile <- calculate_percentile(iris$Sepal.Length, 5.8)
}
