% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PRESS.R
\name{calculate_PRESS}
\alias{calculate_PRESS}
\title{calculate_PRESS}
\usage{
calculate_PRESS(linear_model)
}
\arguments{
\item{linear_model}{The linear model for which to calculate PRESS}
}
\value{
PRESS
}
\description{
Calculates the Predicted Residual Error Sum of Squares for a linear model
}
\examples{
PRESS <- calculate_PRESS(lm(iris$Sepal.Length ~ iris$Sepal.Width, data = iris))
}
