% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_residual_mean_square.R
\name{calculate_residual_mean_square}
\alias{calculate_residual_mean_square}
\title{calculate_residual_mean_square}
\usage{
calculate_residual_mean_square(linear_model)
}
\arguments{
\item{linear_model}{The linear model for which to calculate the residual mean square}
}
\value{
The residual mean square
}
\description{
Calculates the residual mean square of a multiple linear model
}
\examples{
residual_mean_square <- calculate_residual_mean_square(lm(iris$Sepal.Length ~ iris$Sepal.Width, data = iris))
}
