% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMode.R
\name{calculateMode}
\alias{calculateMode}
\title{calculateMode}
\usage{
calculateMode(data_for_sample)
}
\arguments{
\item{data_for_sample}{A vector of data for a sample}
}
\value{
The mode
}
\description{
Calculates the mode given data for a sample
}
\examples{
mode <- calculateMode(iris$Sepal.Length)
}
