% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePercentile.R
\name{calculatePercentile}
\alias{calculatePercentile}
\title{calculatePercentile}
\usage{
calculatePercentile(data_for_sample, value)
}
\arguments{
\item{data_for_sample}{A vector of data for a sample}

\item{value}{A number for which to find the percentile given data for a sample}
}
\value{
A percentile
}
\description{
Calculates the percentile of a value given data for a sample
}
\examples{
percentile <- calculatePercentile(iris$Sepal.Length, 5.8)
}
