% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/construct_confidence_interval_for_population_mean.R
\name{construct_confidence_interval_for_population_mean}
\alias{construct_confidence_interval_for_population_mean}
\title{construct_confidence_interval_for_population_mean}
\usage{
construct_confidence_interval_for_population_mean(
  data_for_sample,
  significance_level
)
}
\arguments{
\item{data_for_sample}{A vector of data for a sample}

\item{significance_level}{A significance level (e.g., 0.05)}
}
\value{
A confidence interval
}
\description{
Constructs a confidence interval for a population mean based on a set of statistics for a sample
}
\examples{
confidence_interval <- construct_confidence_interval_for_population_mean(iris$Sepal.Length, 0.05)
}
