% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/constructConfidenceIntervalForDifferenceBetweenTwoPopulationMeans.R
\name{constructConfidenceIntervalForDifferenceBetweenTwoPopulationMeans}
\alias{constructConfidenceIntervalForDifferenceBetweenTwoPopulationMeans}
\title{constructConfidenceIntervalForDifferenceBetweenTwoPopulationMeans}
\usage{
constructConfidenceIntervalForDifferenceBetweenTwoPopulationMeans(
  data_for_sample_1,
  data_for_sample_2,
  significance_level
)
}
\arguments{
\item{data_for_sample_1}{A vector of numerical data for one sample}

\item{data_for_sample_2}{A vector of numerical data for the other sample}

\item{significance_level}{A significance level (e.g., 0.05)}
}
\value{
A confidence interval
}
\description{
Constructs a confidence interval for the difference between two population means based on two sets of sample statistics
}
\examples{
confidence_interval <- constructConfidenceIntervalForDifferenceBetweenTwoPopulationMeans(iris\%>\%filter(Species=="versicolor")\%>\%pull(Sepal.Length), iris\%>\%filter(Species=="setosa")\%>\%pull(Sepal.Length), 0.05)
}
