% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructConfidenceIntervalForPopulationMean.R
\name{constructConfidenceIntervalForPopulationMean}
\alias{constructConfidenceIntervalForPopulationMean}
\title{constructConfidenceIntervalForPopulationMean}
\usage{
constructConfidenceIntervalForPopulationMean(
  data_for_sample,
  significance_level
)
}
\arguments{
\item{data_for_sample}{A vector of data for a sample}

\item{significance_level}{A significance level (e.g., 0.05)}
}
\value{
A confidence interval
}
\description{
Constructs a confidence interval for a population mean based on a set of statistics for a sample
}
\examples{
confidence_interval <- constructConfidenceIntervalForPopulationMean(iris$Sepal.Length, 0.05)
}
