% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_categorical_vector.R
\name{convert_to_categorical_vector}
\alias{convert_to_categorical_vector}
\title{convert_to_categorical_vector}
\usage{
convert_to_categorical_vector(
  discrete_quantitative_vector,
  vector_of_levels,
  reference_class
)
}
\arguments{
\item{A}{discrete quantitative vector}
}
\value{
A categorical vector
}
\description{
Converts a discrete quantitative vector to a categorical vector
}
\examples{
categorical_vector <- convert_to_categorical_vector(birthwt$race) # MASS::birthwt
}
