% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_frame_for_indicator_variables.R
\name{generate_data_frame_for_indicator_variables}
\alias{generate_data_frame_for_indicator_variables}
\title{generate_data_frame_for_indicator_variables}
\usage{
generate_data_frame_for_indicator_variables(
  factor_of_values_of_categorical_variable
)
}
\arguments{
\item{factor_of_values_of_categorical_variable}{A factor of values of a categorical variable}
}
\value{
A data frame for indicator variables based on a factor of values of a categorical variable
}
\description{
Generates a data frame for indicator variables based on a factor of values of a categorical variable
}
\examples{
data_frame_for_indicator_variables <- generate_data_frame_for_indicator_variables(penguins$species)
}
