% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_index.R
\name{get_column_index}
\alias{get_column_index}
\title{get_column_index}
\usage{
get_column_index(data_frame, column_name)
}
\arguments{
\item{The}{column name}
}
\value{
The column index
}
\description{
Gets the index in a data frame given a column name
}
\examples{
column_index <- get_column_index(data_frame, "column")
}
