% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_of_column_of_data_frame.R
\name{get_index_of_column_of_data_frame}
\alias{get_index_of_column_of_data_frame}
\title{get_index_of_column_of_data_frame}
\usage{
get_index_of_column_of_data_frame(data_frame, column_name)
}
\arguments{
\item{The}{column name}
}
\value{
The column index
}
\description{
Gets the index in a data frame given a column name
}
\examples{
column_index <- get_index_of_column_of_data_frame(data_frame, "column")
}
