% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_of_label_of_vector.R
\name{get_index_of_label_of_vector}
\alias{get_index_of_label_of_vector}
\title{get_index_of_label_of_vector}
\usage{
get_index_of_label_of_vector(vector, label)
}
\arguments{
\item{The}{label}
}
\value{
The index of the label
}
\description{
Gets the index of a label in the vector of names of a vector
}
\examples{
index_of_label <- get_index_of_label_of_vector(vector, "label")
}
