% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_row_index.R
\name{get_row_index}
\alias{get_row_index}
\title{get_row_index}
\usage{
get_row_index(data_frame, row_name)
}
\arguments{
\item{The}{row name}
}
\value{
The row index
}
\description{
Gets the index in a data frame given a row name
}
\examples{
row_index <- get_row_index(data_frame, "row")
}
