% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_K.R
\name{optimize_K}
\alias{optimize_K}
\title{plot_decimal_of_correct_predictions_vs_K}
\usage{
optimize_K(formula, training_data, testing_data)
}
\arguments{
\item{formula}{The formula for the model}

\item{training_data}{A data frame of training data}

\item{testing_data}{A data frame of test data}

\item{type_of_model}{The type of model. An element in the set {"LR", "LDA", "QDA", "KNN"}.}

\item{K}{The parameter of a K Nearest Neighbors model}
}
\value{
summary_of_performance The summary of performance
}
\description{
Generates a summary of performance for a model given a data set for which to predict
}
\examples{
summary_of_performance <- generate_summary_of_performance(type_of_model = "KNN", formula = Direction ~ Lag1 + Lag2, training_data = ISLR2::Weekly_from_1990_to_2008_inclusive, testing_data = ISLR2::Weekly_from_2009_to_2010_inclusive, K = 350)
}
