% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_Box_Cox_Method.R
\name{perform_Box_Cox_Method}
\alias{perform_Box_Cox_Method}
\title{perform_Box_Cox_Method}
\usage{
perform_Box_Cox_Method(
  linear_model,
  vector_of_values_of_lambda,
  whether_to_plot
)
}
\arguments{
\item{linear_model}{A linear model}

\item{vector_of_values_of_lambda}{Defaults to seq(-2, 2, 0.1)}

\item{whether_to_plot}{Defaults to true}
}
\value{
result_of_Box_Cox_Method
}
\description{
Performs the Box-Cox Method
}
\examples{
result_of_Box_Cox_Method <- perform_Box_Cox_Method(linear_model)
}
