% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.analysis_of_variance.R
\name{print.analysis_of_variance}
\alias{print.analysis_of_variance}
\title{print.analysis_of_variance}
\usage{
\method{print}{analysis_of_variance}(analysis)
}
\arguments{
\item{analysis}{The analysis to print}
}
\description{
Adds to the generic function print the analysis_of_variance method print.analysis_of_variance
}
\examples{
print(analysis_of_variance)

}
