% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeVariance.R
\name{print.AnalysisOfVariance}
\alias{print.AnalysisOfVariance}
\title{print.linearModelSummary}
\usage{
\method{print}{AnalysisOfVariance}(the_analysis_of_variance)
}
\arguments{
\item{the_summary}{The linear-model summary to print}
}
\description{
Adds to the generic function print the LinearModelSummary method print.LinearModelSummary
}
\examples{
print(the_linear_model_summary)
}
