% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNullHypothesisInvolvingProportion.R
\name{print.HypothesisTestResult}
\alias{print.HypothesisTestResult}
\title{print.HypothesisTestResult}
\usage{
\method{print}{HypothesisTestResult}(the_hypothesis_test_result)
}
\arguments{
\item{the_summary}{The hypothesis-test-result to print}
}
\description{
Adds to the generic function print the HypothesisTestResult method print.HypothesisTestResult
}
\examples{
print(the_hypothesis_test_result)
}
