% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_model.R
\name{print.linear_model_summary}
\alias{print.linear_model_summary}
\title{print.linear_model_summary}
\usage{
\method{print}{linear_model_summary}(linear_model_summary)
}
\arguments{
\item{summary_output}{The linear-model summary to print}
}
\description{
Adds to the generic function print the linear_model_summary method print.linear_model_summary
}
\examples{
print(linear_model_summary)
}
