% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeLinearModel.R
\name{print.LinearModelSummary}
\alias{print.LinearModelSummary}
\title{print.linearModelSummary}
\usage{
\method{print}{LinearModelSummary}(the_linear_model_summary)
}
\arguments{
\item{the_summary}{The linear-model summary to print}
}
\description{
Adds to the generic function print the LinearModelSummary method print.LinearModelSummary
}
\examples{
print(the_linear_model_summary)
}
