% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_data_for_ROC_curve.R
\name{provide_data_for_ROC_curve}
\alias{provide_data_for_ROC_curve}
\title{provide_data_for_ROC_curve}
\usage{
provide_data_for_ROC_curve(
  vector_of_actual_indicators,
  vector_of_predicted_probabilities
)
}
\arguments{
\item{vector_of_actual_indicators}{The vector of actual indicators in the set {0, 1}}

\item{vector_of_predicted_probabilities}{The vector of predicted probabilities in the range [0, 1]}
}
\value{
data_frame_of_thresholds_TPRs_and_FPRs
}
\description{
Provides data for ROC curve
}
\examples{
data_frame_of_thresholds_TPRs_and_FPRs <- provide_data_for_ROC_curve(vector_of_actual_indicators, vector_of_predicted_probabilities)
}
