% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_performance_metrics.R
\name{provide_performance_metrics}
\alias{provide_performance_metrics}
\title{provide_performance_metrics}
\usage{
provide_performance_metrics(
  vector_of_actual_indicators,
  vector_of_predicted_probabilities
)
}
\arguments{
\item{vector_of_actual_indicators}{The vector of actual indicators in the set {0, 1}}

\item{vector_of_predicted_probabilities}{The vector of predicted probabilities in the range [0, 1]}
}
\value{
data_frame_of_performance_metrics
}
\description{
Provides performance metrics
}
\examples{
data_frame_of_performance_metrics <- provide_performance_metrics(vector_of_actual_indicators, vector_of_predicted_probabilities)
}
