% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/provide_vector_of_values_of_performance_metric.R
\name{provide_vector_of_values_of_performance_metric}
\alias{provide_vector_of_values_of_performance_metric}
\title{provide_vector_of_values_of_performance_metric}
\usage{
provide_vector_of_values_of_performance_metric(
  vector_of_actual_indicators,
  vector_of_predicted_probabilities,
  performance_metric
)
}
\arguments{
\item{vector_of_actual_indicators}{The vector of actual indicators in the set {0, 1}}

\item{vector_of_predicted_probabilities}{The vector of predicted probabilities in the range [0, 1]}
}
\value{
vector_of_values_of_performance_metric
}
\description{
Provides vector of values of performance metric
}
\examples{
vector_of_values_of_performance_metric <- provide_vector_of_values_of_performance_metric(vector_of_actual_indicators, vector_of_predicted_probabilities)
}
