% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_between_0_and_1.R
\name{scale_to_between_0_and_1}
\alias{scale_to_between_0_and_1}
\title{scale_to_between_0_and_1}
\usage{
scale_to_between_0_and_1(the_vector)
}
\arguments{
\item{the_vector}{The vector to scale to between 0 and 1}
}
\value{
scaled_vector The scaled vector
}
\description{
Scales to between 0 and 1
}
\examples{
scaled_vector <- scale_to_between_0_and_1(the_vector)
}
