% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_set_into_training_and_testing_sets.R
\name{split_data_set_into_training_and_testing_data}
\alias{split_data_set_into_training_and_testing_data}
\title{split_data_set_into_training_and_testing_sets}
\usage{
split_data_set_into_training_and_testing_data(
  data_frame,
  proportion_of_training_data
)
}
\arguments{
\item{data_frame}{The data set to split into training and testing sets}

\item{proportion_of_training_data}{The proportion of training data}
}
\value{
split_data The split data
}
\description{
Splits data set into training and testing sets
}
\examples{
split_data <- split_data_set_into_training_and_testing_sets(data_frame = Default, proportion_of_training_data = 0.5)
}
