% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_model.R
\name{summarize_linear_model}
\alias{summarize_linear_model}
\title{summarize_linear_model}
\usage{
summarize_linear_model(linear_model)
}
\arguments{
\item{linear_model}{The linear model to summarize}
}
\value{
The summary for the linear model
}
\description{
Summarizes a linear model
}
\examples{
summary_output <- summarize_linear_model(lm(iris$Sepal.Length ~ iris$Sepal.Width, data = iris))
}
