% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/summarize_performance_of_cross_validated_classifiers.R
\name{summarize_performance_of_cross_validated_classifiers}
\alias{summarize_performance_of_cross_validated_classifiers}
\title{summarize_performance_of_cross_validated_classifiers}
\usage{
summarize_performance_of_cross_validated_classifiers(
  type_of_model,
  formula,
  data_frame,
  optimal_lambda = NULL
)
}
\arguments{
\item{type_of_model}{The type of classifier for which to summarize performance of cross validated classifiers}

\item{formula}{The formula of the classifier for which to summarize performance of cross validated classifiers}

\item{data_frame}{The data frame for which to summarize performance of trained and cross validated classifiers}

\item{optimal_lambda}{The optimal value of lambda for which to summarize performance of trained and cross validated logistic ridge regression models}
}
\value{
summary_of_performance The summary of performance of cross-validated classifiers
}
\description{
Summarizes performance of cross validated classifiers
}
\examples{
summary_of_performance <- summarize_performance_of_cross_validated_classifiers("Logistic Regression", Indicator ~ Red + Green + Blue, data_frame = data_frame_of_indicators_and_pixels, optimal_lambda = 0.001)
}
