% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/summarize_performance_of_cross_validated_models.R
\name{summarize_performance_of_cross_validated_models}
\alias{summarize_performance_of_cross_validated_models}
\title{summarize_performance_of_cross_validated_models}
\usage{
summarize_performance_of_cross_validated_models(
  type_of_model,
  formula,
  data_frame,
  number_of_folds,
  threshold = 0.5
)
}
\arguments{
\item{type_of_model}{The type of model. An element in the set {"Polynomial Regression", Logistic Regression"}.}
}
\value{
summary_of_performance_of_cross_validated_models The summary of performance of cross-validated models
}
\description{
Summarizes performance of cross-validated models
}
\examples{
summary_of_performance_of_cross_validated_models <- summarize_performance_of_cross_validated_models(type_of_model = "LR", formula = Direction ~ Lag1 + Lag2, data_frame = ISLR2::Weekly)
}
