% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/summarize_performance_of_cross_validated_models_using_dplyr.R
\name{summarize_performance_of_cross_validated_models_using_dplyr}
\alias{summarize_performance_of_cross_validated_models_using_dplyr}
\title{summarize_performance_of_cross_validated_models_using_dplyr}
\usage{
summarize_performance_of_cross_validated_models_using_dplyr(
  type_of_model,
  formula,
  data_frame
)
}
\arguments{
\item{formula}{The formula of the logistic regression model for which to summarize performance of cross validated models using dplyr}

\item{data_frame}{The data frame on which to train the logistic regression for which to summarize performance of cross validated models using dplyr}
}
\value{
summary_of_performance_of_cross_validated_models_using_dplyr The summary of performance of cross-validated models using dplyr
}
\description{
Summarizes performance of cross validated models using dplyr
}
\examples{
summary_of_performance_of_cross_validated_models_using_dplyr <- summarize_performance_of_cross_validated_models_using_dplyr(Indicator ~ Red + Green + Blue, data_frame = data_frame_of_indicators_and_pixels)
}
