% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_performance_of_one_model.R
\name{summarize_performance_of_one_model}
\alias{summarize_performance_of_one_model}
\title{summarize_performance_of_one_model}
\usage{
summarize_performance_of_one_model(
  type_of_model,
  formula,
  training_data,
  test_data,
  K = 1
)
}
\arguments{
\item{type_of_model}{The type of model. An element in the set {"Polynomial Regression", "Logistic Regression", "LDA", "QDA", "KNN"}.}

\item{formula}{The formula for the model}

\item{training_data}{A data frame of training data}

\item{test_data}{A data frame of test data}

\item{K}{The parameter of a K Nearest Neighbors model}
}
\value{
summary_of_performance_of_one_model The summary of performance
}
\description{
Generates a summary of performance for one model given a data set for which to predict
}
\examples{
summary_of_performance_of_one_model <- summarize_performance_of_one_model(type_of_model = "KNN", formula = Direction ~ Lag1 + Lag2, training_data = ISLR2::Weekly_from_1990_to_2008_inclusive, test_data = ISLR2::Weekly_from_2009_to_2010_inclusive, K = 350)
}
