% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeLinearModel.R
\name{summarizeLinearModel}
\alias{summarizeLinearModel}
\title{summarizeLinearModel}
\usage{
summarizeLinearModel(linear_model)
}
\arguments{
\item{linear_model}{The linear model for which to generate a summary}
}
\value{
The summary for the linear model
}
\description{
Summarizes a linear model
}
\examples{
the_summary <- summarizeLinearModel(the_linear_model)
}
