% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/test_null_hypothesis_involving_MLR_coefficients.R
\name{test_null_hypothesis_involving_MLR_coefficients}
\alias{test_null_hypothesis_involving_MLR_coefficients}
\title{test_null_hypothesis_involving_MLR_coefficients}
\usage{
test_null_hypothesis_involving_MLR_coefficients(
  linear_model,
  significance_level
)
}
\arguments{
\item{A}{significance level}
}
\value{
test_result A message rejecting or not rejecting the null hypothesis, and supporting or not supporting the alternate hypothesis
}
\description{
Tests a null hypothesis involving the coefficients of a Multiple Linear Regression model
}
\examples{
test_result <- test_null_hypothesis_involving_MLR_coefficients(lm(iris$Sepal.Length ~ iris$Sepal.Width, data = iris), 0.05)
}
