% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_null_hypothesis_involving_slope.R
\name{test_null_hypothesis_involving_slope}
\alias{test_null_hypothesis_involving_slope}
\title{test_null_hypothesis_involving_slope}
\usage{
test_null_hypothesis_involving_slope(linear_model, significance_level)
}
\arguments{
\item{A}{significance level}
}
\value{
test_result A message rejecting or not rejecting the null hypothesis, and supporting or not supporting the alternate hypothesis
}
\description{
Tests a null hypothesis involving a slope
}
\examples{
test_result <- test_null_hypothesis_involving_slope(0.05)
}
