% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNullHypothesisInvolvingProportion.R
\name{testNullHypothesisInvolvingProportion}
\alias{testNullHypothesisInvolvingProportion}
\title{testNullHypothesisInvolvingProportion}
\usage{
testNullHypothesisInvolvingProportion(
  population_proportion,
  sample_proportion,
  number_of_data_for_sample,
  comparator,
  significance_level
)
}
\arguments{
\item{population_proportion}{The population proportion in the null hypothesis}

\item{sample_proportion}{A sample proportion}

\item{comparator}{"=", "<=", or ">="}
}
\value{
testResult A message rejecting or not rejecting the null hypothesis, and supporting or not supporting the alternate hypothesis
}
\description{
Tests a null hypothesis involving a proportion
}
\examples{
testResult <- testNullHypothesisInvolvingProportion(0.758, 0.505, 467, ">=", 0.05)
}
