% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_critical_value_Fc.R
\name{calculate_critical_value_Fc}
\alias{calculate_critical_value_Fc}
\title{calculate_critical_value_Fc}
\usage{
calculate_critical_value_Fc(
  significance_level,
  regression_degrees_of_freedom,
  residual_degrees_of_freedom
)
}
\arguments{
\item{The}{residual degrees of freedom}
}
\value{
The critical value F
}
\description{
Calculates critical value F of an F distribution with provided numbers of regression and residual degrees of freedom with provided significance level
}
\examples{
critical_value_F <- calculate_critical_value_Fc(0.05, 2, 52)
}
