% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hypothesis_test_result.R
\name{print.hypothesis_test_result}
\alias{print.hypothesis_test_result}
\title{print.hypothesis_test_result}
\usage{
\method{print}{hypothesis_test_result}(hypothesis_test_result)
}
\arguments{
\item{the_summary}{The hypothesis-test result to print}
}
\description{
Adds to the generic function print the hypothesis_test_result method print.hypothesis_test_result
}
\examples{
print(hypothesis_test_result)
}
